Move log functions from util.h to their own headers.
Also adds file:line# to all log statements when trace is enabled.
