+++ src/render/djgpp/renderer_svga.c
@@ -30,8 +30,8 @@
 #include "../render.h"
 #include "../render_layer.h"
 #include "../renderers.h"
-#include "../../util.h"
 
+#include "../../platform/log.h"
 #include "../../platform/djgpp/platform_djgpp.h"
 
 struct svga_render_data
