+++ src/util.c
@@ -458,102 +458,6 @@ const char *mzx_res_get_by_id(enum resource_id id)
   return mzx_res[id].path;
 }
 
-#ifdef CONFIG_STDIO_REDIRECT
-
-FILE *mzxout_h = NULL;
-FILE *mzxerr_h = NULL;
-
-/**
- * Some platforms may not be able to display console output without extra work.
- * On these platforms, open stdout/stderr replacement files instead. The log
- * macros and any other references to `mzxout` and `mzxerr` will use these
- * files instead of stdio.
- *
- * Previously, this was implemented using `freopen` on stdout/stderr, but
- * doing this in some console SDKs does not work correctly (NDS, PS Vita).
- */
-boolean redirect_stdio_init(const char *base_path, boolean require_conf)
-{
-  char dest_path[MAX_PATH];
-  size_t dest_len;
-  FILE *fp_wr;
-  uint64_t t;
-
-  if(!base_path)
-    return false;
-
-  dest_len = path_clean_copy(dest_path, MAX_PATH - 10, base_path);
-
-  if(require_conf)
-  {
-    // If the config file is required, attempt to stat it.
-    struct stat stat_info;
-
-    path_append(dest_path, MAX_PATH, "config.txt");
-    if(vstat(dest_path, &stat_info))
-      return false;
-    dest_path[dest_len] = '\0';
-  }
-
-  // Clean up existing handles from a previous attempt.
-  redirect_stdio_exit();
-
-  // Test directory for write access.
-  path_append(dest_path, MAX_PATH, "stdout.txt");
-  fp_wr = fopen_unsafe(dest_path, "w");
-  if(!fp_wr)
-  {
-    fprintf(stdout, "Failed to redirect stdout\n");
-    fflush(stdout);
-    return false;
-  }
-
-  t = (uint64_t)time(NULL);
-
-  // Redirect mzxout to stdout.txt.
-  fprintf(stdout, "Redirecting logs to '%s'...\n", dest_path);
-  fflush(stdout);
-  mzxout_h = fp_wr;
-  fprintf(mzxout, "MegaZeux: Logging to '%s' (%" PRIu64 ")\n", dest_path, t);
-  fflush(mzxout);
-
-  // Redirect mzxerr to stderr.txt.
-  dest_path[dest_len] = '\0';
-  path_append(dest_path, MAX_PATH, "stderr.txt");
-  fp_wr = fopen_unsafe(dest_path, "w");
-  if(!fp_wr)
-  {
-    fprintf(stderr, "Failed to redirect stderr\n");
-    fflush(stderr);
-    return false;
-  }
-
-  fprintf(stderr, "Redirecting logs to '%s'...\n", dest_path);
-  fflush(stderr);
-  mzxerr_h = fp_wr;
-  fprintf(mzxerr, "MegaZeux: Logging to '%s' (%" PRIu64 ")\n", dest_path, t);
-  fflush(mzxerr);
-
-  return true;
-}
-
-void redirect_stdio_exit(void)
-{
-  if(mzxout_h)
-  {
-    fclose(mzxout_h);
-    mzxout_h = NULL;
-  }
-
-  if(mzxerr_h)
-  {
-    fclose(mzxerr_h);
-    mzxerr_h = NULL;
-  }
-}
-
-#endif /* CONFIG_STDIO_REDIRECT */
-
 // Random function, returns an integer [0-range)
 
 static uint64_t rng_state;
